+++ src/components/pagination/type/index.ts
@@ -1,0 +1,12 @@
+
+
+// 分页接口
+export interface PagesType {
+    total: number
+    size: number
+}
+
+export interface TableFooterType {
+    currentPage: number
+    pages: PagesType
+}
